Move custom assertions to a separate file
